redux로 만들기

+++ front/components/comment/CommentItem.jsx
@@ -1,15 +1,21 @@
 import { useState } from "react";
 import Link from "next/link";
-import { useDispatch } from "react-redux";
-import { DeleteComment } from "../../reducers/comment";
-import { destroyComment } from "../api/Comment";
+import { useDispatch, useSelector } from "react-redux";
+import { GetReplys,ClearReplys,DeleteComment,DeleteReply } from "../../reducers/article";
+import { destroyComment,showComment } from "../api/Comment";
 import UpdateForm from "./CommentUpdateForm";
 import CommentForm from "./CommentForm";
 import CommentList from "./CommentList";
 import styled from "styled-components";
 import { LikeBtn } from "../board/LikeBtn";
+import { KAKAO_AUTH_URL } from "../../components/api/OAuth";
+import Router from "next/router";
 
 const StyledCommentItem = styled.div`
+  width: 100%;
+  height: auto;
+  overflow: hidden;
+  margin: 10px 0;
 & > .image {
     float: left;
     width: 50px;
@@ -22,7 +28,7 @@ const StyledCommentItem = styled.div`
 
 &>ul{
   float: left;
-    width: 90%;
+    width: 85%;
     height: auto;
     overflow: hidden;
 }
@@ -50,9 +56,12 @@ const CommentItem = ({
   isWriter,
 }) => {
   const dispatch = useDispatch();
+  const article = useSelector(state=>state.article)
+  const {nickname,IsLogin} = useSelector((state)=>state.user)
   const [update, setUpdate] = useState(false);
   const [createReply, setCreateReply] = useState(false);
   const [showReply, setShowReply] = useState(false);
+  const [skip, setSkip] = useState(0);
 
   const handleUpdate = (value) => {
     setUpdate(value);
@@ -62,19 +71,50 @@ const CommentItem = ({
     setCreateReply(value);
   };
 
-  const handleShow = (value) => {
-    setTimeout(setShowReply(value), 1000);
+  const handleShow = async(value) => {
+    setTimeout(setShowReply(value));
+    // if(value){
+      //답글불러오기
+      const data = {
+        board_id:article.board_id,
+        skip:skip,
+        root:comment_id,
+        type:'old'
+      }
+      const result = await showComment(data);
+      if(result.length>0){
+        dispatch(GetReplys(result))
+        setSkip(skip+10)
+      }
+    // }else{
+    //   dispatch(ClearReplys(comment_id))
+    //   setSkip(0)
+    // }
   };
 
+
+
+
   const handleDelete = async () => {
-    const answer = confirm("삭제하시겠습니까?");
-    if (answer) {
-      const data = { comment_id: comment_id, writer: writer };
-      const result = await destroyComment(data);
-      if (result.success) {
-        dispatch(DeleteComment(comment_id));
-      } else {
-        alert(result.error);
+    if(IsLogin){
+
+      const answer = confirm("삭제하시겠습니까?");
+      if (answer) {
+        const data = { comment_id: comment_id, writer: writer };
+        const result = await destroyComment(data);
+        if (result.success) {
+         if(root==0){
+           dispatch(DeleteComment(comment_id));
+         }else{
+          dispatch(DeleteReply({comment_id,root}));
+         }
+        } else {
+          alert(result.error);
+        }
+      }
+    }else{
+      if(confirm('로그인하시겠습니까?')){
+        Router.push(`${KAKAO_AUTH_URL}`)
       }
     }
   };
@@ -111,7 +151,7 @@ const CommentItem = ({
 
           <li>
             
-              <LikeBtn liked={liked} disliked={disLiked} isLike={isLike} type={'clike'} id={comment_id} />
+              <LikeBtn liked={liked} disliked={disliked} isLike={isLike} type={'clike'} id={comment_id} />
             
             {createReply === false ? (
               <span
@@ -132,10 +172,10 @@ const CommentItem = ({
             )}
           </li>
           {createReply && (
-            <CommentForm root={comment_id} handleCreate={handleCreate} board_id={board_id}/>
+            <CommentForm root={comment_id} handleCreate={handleCreate}/>
           )}
-          <li>
             {parseInt(reply_cnt) > 0 && (
+          <li>
               <>
                 {showReply === false ? (
                   <span
@@ -155,12 +195,12 @@ const CommentItem = ({
                   </span>
                 )}
               </>
-            )}
           </li>
+            )}
           {showReply && (
             <li>
               <ul>
-                <CommentList root={comment_id} isReply={true} board_id={board_id}/>
+                <CommentList root={comment_id}/>
               </ul>
             </li>
           )}
